+++ include/ColorSensor.h
@@ -11,8 +11,6 @@ class ColorSensor{
         void calibra_sensor_frontal(int num);
         void calibra_sensor_inferior(int num);
         int identify_color();
-        int identify_color2();
-        int identify_color3();
         void read_limit_values(int num);
         void calibra_cor_inferior(char cor[20], int count);
         void calibra_cor_frontal(char cor[20], int count);
@@ -84,7 +82,6 @@ class ColorSensor{
         int lim_white_b;
 
 
-    
 
 };
 #endif
